ResourceContains has to iterate over all elements
The ResourceContains method was checking only if first resource were
containing the requested value.

For example, if first element did not have the value but the second did,
ResourceContains returned false, because it did not check the second element.

The loop is moved outside to easily unit test ResourceContains: otherwise it
would require to mock kubernetes client and its methods.

As the new unit test is very similar to the one for ObjectContains, we removed
the later and extended metadata field in mock object values to cover all
ObjectContains case too.